requirements.txt
keras==2.9.0rc0
##numpy==1.22.3
Pillow==9.0.1
scikit-image==0.19.2
streamlit==1.7.0
face_detection==0.2.2
opencv-python-headless==4.5.5.64
streamlit-drawable-canvas==0.8.0
click==8.0.4
tensorflow==2.9.0
click==8.0.4
torch>=1.7
numpy<1.21  # numba requires numpy<1.21,>=1.17
opencv-python
torchvision
scipy
tqdm
basicsr>=1.3.4.0
facexlib>=0.2.0.3
lmdb
pyyaml
tb-nightly
yapf

